+++ lib/src/Common/Extensions/version_path_part.dart
@@ -2,23 +2,23 @@ part of softhai.bulls_eye.Common;
 
 class Version extends Variable {
 
-  static String routePartVariableVersion = "Version";
+  static String pathPartVariableVersion = "Version";
 
-  Version([Map<String,dynamic> extensions]) : super(routePartVariableVersion, isOptional: false, extensions: extensions) 
+  Version([Map<String,dynamic> extensions]) : super(pathPartVariableVersion, isOptional: false, extensions: extensions) 
   {
     RegisterParseFunc();
   }
   
-  void CheckUsage(Url currentRoute) 
+  void CheckUsage(Url currentUrl) 
   {
-    if(currentRoute.routeParts.where((part) => part is Version).length > 1) throw new MultipleVersionException(currentRoute);
+    if(currentUrl.pathParts.where((part) => part is Version).length > 1) throw new MultipleVersionException(currentUrl);
   }
   
-  static Variable _TryParse(String rawStr, String cleanStr, bool isOptional, UrlDefConfig currentConfig, Url currentRoute) {
+  static Variable _TryParse(String rawStr, String cleanStr, bool isOptional, UrlDefConfig currentConfig, Url currentUrl) {
     
-    if(cleanStr == routePartVariableVersion && !isOptional)
+    if(cleanStr == pathPartVariableVersion && !isOptional)
     {
-      if(currentRoute.routeParts.where((part) => part is Version).length > 0) throw new MultipleVersionException(currentRoute);
+      if(currentUrl.pathParts.where((part) => part is Version).length > 0) throw new MultipleVersionException(currentUrl);
 
       return new Version();
     }
@@ -38,11 +38,11 @@ class Version extends Variable {
 
 class MultipleVersionException implements Exception {
 
-  Url _routeDefinition;
+  Url _urlDefinition;
   
-  MultipleVersionException(this._routeDefinition);
+  MultipleVersionException(this._urlDefinition);
   
   String toString() {
-    return "Multible Version Definition in Route: " + this._routeDefinition.name;
+    return "Multible Version Definition in Route: " + this._urlDefinition.name;
   }
 }